05-07-24

+++ webappd/webappd/Models/CombinedSpices.cs
@@ -11,6 +11,6 @@ namespace webappd.Models
         public string ImageName { get; set; }
         public string ImagePath { get; set; }
         public string Description { get; set; }
-        public int Price { get; set; }
+        public Decimal Price { get; set; }
     }
 }
